Added conftest file
