+++ src/tests/conftest.py
@@ -1,0 +1,27 @@
+import pytest
+import json
+import uuid
+
+
+@pytest.fixture()
+def request_headers() -> dict:
+    return {
+        "content-type": "application/vnd.api+json",
+        "accept": "application/vnd.api+json",
+        "x-correlation-id": str(uuid.uuid4()),
+    }
+
+
+@pytest.fixture()
+def request_body() -> json:
+    return {
+        "data": {
+            "type": "Message",
+            "attributes": {
+                "messageReference": str(uuid.uuid4()),
+                "recipient": {"nhsNumber": "9990548609", "dateOfBirth": "1932-01-06"},
+                "originator": {"odsCode": "X26"},
+                "personalisation": {"custom": "value"},
+            },
+        }
+    }
