+++ src/tests/test_base_api_client.py
@@ -1,35 +1,7 @@
-import pytest
-import os
-import uuid
 import json
 from services.BaseAPIClient import BaseAPIClient
 from services.AuthManager import AuthManager
-from unittest.mock import patch
 from dotenv import dotenv_values
-
-
-@pytest.fixture()
-def request_headers() -> dict:
-    return {
-        "content-type": "application/vnd.api+json",
-        "accept": "application/vnd.api+json",
-        "x-correlation-id": str(uuid.uuid4()),
-    }
-
-
-@pytest.fixture()
-def request_body() -> json:
-    return {
-        "data": {
-            "type": "Message",
-            "attributes": {
-                "messageReference": str(uuid.uuid4()),
-                "recipient": {"nhsNumber": "9990548609", "dateOfBirth": "1932-01-06"},
-                "originator": {"odsCode": "X26"},
-                "personalisation": {"custom": "value"},
-            },
-        }
-    }
 
 
 class TestBaseAPIClient:
